Exclude compare tasks from the concurrency limit
Previously, capture and compare were treated as a single task on the queue. A new
capture/compare task could only be started when a previous task had completed.
This change makes it so that captures can run independently of the queue, which
allows more captures to happen concurrently. In my tests, this resulted in a 5-15%
speedup of the overall test command.